move dump_task_info_table
